tags: add list and resize-able content
