+++ plugins/chrome/src/components/tags/Index.svelte
@@ -1,147 +1,30 @@
 <script>
 import { onMount } from 'svelte';
-import { tags } from './stores.js';
-import { states } from '../button/states.js';
-import  Chevron from '../button/Chevron.svelte';
-import BStatic from '../box/BStatic.svelte';
-import BHeader from '../box/BHeader.svelte';
-import BTable from '../box/BTable.svelte';
-import Button from './Button.svelte';
-import Tags1 from './Tags1_.svelte'; 
-import Tags2 from './Tags2_.svelte'; 
-import Tags3 from './Tags3_.svelte'; 
-import Urls from './Urls.svelte';
+// import { source } from './stores.js';
 
-export let top = "23";
+import VBox2 from '../box/VBox2.svelte';
+import Content from './Content.svelte';
+import List from './List.svelte';
 
-let block = true;
-let cols = 3;
+let left = 165;
+const top = '47';
+const title = '-Tag(s)-' 
+const id = 'routeLeft';
 
 onMount(async () => {
-  console.warn('onMount tags/index');
+
 });
 
-window.mitm.files.getRoute_events.tagsTable = () => {
-  // window.ws__send('getRoute', '', routeHandler);
-  console.log('events.tagsTable...');
-  const {__tag1, __tag2, __tag3} = window.mitm;
-  const {filterUrl, uniq} = $tags;
-  const tgroup = {};
-  for (let ns in __tag2) {
-    const tsks = __tag2[ns]
-    for (let task in tsks) {
-      const [k,v] = task.split(':');
-      if (v && k!=='url') {
-        tgroup[v] = true
-      }
-    }
-  } 
-  tags.set({
-    filterUrl,
-    __tag1,
-    __tag2,
-    __tag3,
-    tgroup,
-    uniq
-  })
+function dragend({detail}) {
+
 }
-function oneClick(e) {
-  const {_cols} = e.target.dataset;
-  cols = +_cols;
-}
-function handleMessage(event) {
-  const {all, name} = event.detail
-  let q
-  if (name==='state2') {
-    all.chevron = all.state2 ? '[<<]' : '[<<]'
-    all.state3 = {} // feat: auto collapsed between tag2 & tag3
-    q = '.t3'
-  } else if (name==='state3') {
-    all.chevron = !all.state3 ? '[>>]' : '[>>]'
-    all.state2 = {} // feat: auto collapsed between tag2 & tag3
-    q = '.t2'
-  }
-  const nodes = document.querySelectorAll(`${q} details[open]`)
-  nodes.forEach(node => node.removeAttribute('open'))  
-  states.set(all)
+
+function onChange(e) {
+
 }
 </script>
 
-<Button/>
-<div class="vbox">
-  <details open="true">
-    <summary>Enable / Disable Tags</summary>
-    <div class="vbox-1">
-      <BStatic {top} {block}>
-        <BHeader>-Tags-
-          <button data-_cols=3 on:click="{oneClick}">[full]</button>
-          <button data-_cols=2 on:click="{oneClick}">[two]</button>
-          <button data-_cols=1 on:click="{oneClick}">[one]</button>
-          <Chevron/>
-        </BHeader>
-        <BTable>
-          <tr class="set-tags">
-            <Tags1 {cols}/>
-            <Tags2 {cols} on:message={handleMessage}/>
-            <Tags3 {cols} on:message={handleMessage}/>
-          </tr>
-        </BTable>
-      </BStatic>  
-    </div>
-  </details>
-  <details class="urls">
-    {@html '<style id="urls"></style>'}
-    <summary>Effected Url(s)</summary>
-    <Urls/>
-  </details>
-</div>
-
-<style>
-.vbox {
-  flex: auto;
-  display: flex;
-  position: relative;
-  flex-direction: column;
-  height: calc(100vh - 23px);
-}
-.vbox-1 {
-  margin-bottom: 10px;
-}
-details, summary {
-  outline: none;
-}
-summary {
-  border: none;
-  cursor: pointer;
-  font-size: 13px;
-  padding-left: 5px;
-  background: #fdaaaa;
-}
-.urls summary:hover {
-  background-color: #f1f6fbbd;
-}
-.urls {
-  height: 100%;
-  display: flex;
-  overflow: auto;
-  flex-direction: column;
-}
-.urls summary {
-  position: sticky;
-  background: white;
-  top: 0px;
-}
-button {
-  border: 0;
-  cursor: pointer;
-  color: #002aff;
-  margin-top: -5px;
-  margin-right: -5px;
-  vertical-align: 0.6px;
-  background: transparent;
-  padding: 2px 1px 1px 1px;
-  font-family: Consolas, Lucida Console, Courier New, monospace;
-  font-weight: 700;
-  font-size: 10px;
-}
-</style>
+<Content {onChange}/>
+<!-- <VBox2 {title} {top} {left} {dragend} {List} props={{onChange}}>
+  <Content {onChange}/>
+</VBox2> -->
